chore: move all support file to legacy target
 - prepare split into libraries
 - remove ezTime dependency to get bazel working
