+++ BUILD
@@ -3,3 +3,14 @@ load(
     "platformio_library",
     "platformio_project",
 )
+
+platformio_project(
+    name = "vario",
+    src = "main.cc",
+    board = "nodemcuv2",
+    framework = "arduino",
+    platform = "espressif8266",
+    deps = [
+        "//legacy"
+    ]
+)
