+++ legacy/Beeper.cpp
@@ -1,4 +1,4 @@
-#include "./Beeper.h"
+#include "Beeper.h"
 #include <queue>
 
 Beeper::Beeper()
