+++ legacy/Vario.h
