+++ legacy/legacy.h
@@ -1,0 +1,8 @@
+#ifndef LEGACY_LEGACY_H
+#define LEGACY_LEGACY_H
+
+#include "Blinker.h"
+#include "Beeper.h"
+#include "Vario.h"
+
+#endif //LEGACY_LEGACY_H
